Урок 7
Примечание: Для задания 2 потоки просто излишни. Мало того, что они буквально ЖРУТ память, так еще и для такого маленького задания они совсем не нужны. Но потоком я воспользовался в первом задании для патруля - там он уместен. И еще - я доточил свою реализацию очереди команд и переделал установку CancellationTokenManager через Zenject.

1) Смотреть файлы папки Assets\Scripts\Core\CommandExecutors\PatrolCommandExecutor

2) В папках Assets\Scripts\UserControlSystem классы FactionCounterModel, FactionCounterPresenter и FactionCounterView. через абстракцию с моделью работает класс FactionMember.

+++ Assets/Scripts/Abstractions/IHealthHolder.cs
@@ -3,8 +3,8 @@
 
     #region Properties
 
-    float Health { get; }
-    float MaxHealth { get; }
+    int Health { get; }
+    int MaxHealth { get; }
 
     #endregion
